Transactions配下にあるファイルを分割
