+++ LogoREGIUI/Sources/Presentation/Transactions/StartTransactionView.swift
@@ -1,9 +1,4 @@
-//
-//  StartTransactionView.swift
-//  cafelogos-pos
-//
-//  Created by Owner on 2023/09/23.
-//
+// レジ開け画面
 
 import SwiftUI
 
@@ -41,47 +36,6 @@ struct StartTransactionView: View {
         }
     }
 }
-
-struct multi2Button<Destination>: View where Destination : View {
-    var title: String
-    var subtitle: String?
-    var bgColor: Color
-    var fgColor: Color
-    var destination: () -> Destination
-    
-    public init(title: String, subtitle: String?, bgColor: Color, fgColor: Color, destination: @escaping () -> Destination) {
-        self.title = title
-        self.subtitle = subtitle
-        self.bgColor = bgColor
-        self.fgColor = fgColor
-        self.destination = destination
-    }
-    
-    var body: some View {
-        NavigationLink(destination: destination()) {
-            VStack(spacing: 0) {
-                Text(title)
-                    .font(.title)
-                    .fontWeight(.semibold)
-                    .lineLimit(0)
-                if let subtitle = subtitle {
-                    Text(subtitle)
-                        .font(.title)
-                }
-            }
-            .foregroundColor(fgColor)
-            .frame(maxWidth: .infinity, minHeight: 50, alignment: .center)
-            .clipped()
-            .padding(.vertical, 20)
-            .background {
-                RoundedRectangle(cornerRadius: 8, style: .continuous)
-                    .fill(bgColor)
-            }
-            .padding(.bottom)
-        }
-    }
-}
-
 
 struct StartTransactionView_Previews: PreviewProvider {
     static var previews: some View {
